% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combData.R
\name{combData}
\alias{combData}
\title{A combData function}
\usage{
combData(deltaCt, ND, sbsBgn, gene, delNondet = TRUE)
}
\arguments{
\item{deltaCt}{= delta Ct values;}

\item{ND}{= vector of nondetect status;}

\item{sbsBgn}{= the first position of different treatments;}

\item{gene}{= gene names;}

\item{delNondet}{= an indicator whether or not we delete genes
that are unexpressed in any of the samples, default = TRUE}
}
\description{
this is a function that makes a dataset for one set
of genes (one plate)
}

