% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonGenesCorPlot.R
\name{CommonGenesCorPlot}
\alias{CommonGenesCorPlot}
\title{A CommonGenesCorPlot function}
\usage{
CommonGenesCorPlot(object, seqGeneSum, extrVal = FALSE)
}
\arguments{
\item{object}{= qPCR object;}

\item{seqGeneSum}{= pre processed RNAseq data, for each sample has mean of the log2(CPM).}

\item{extrVal}{= logical indicator , if TRUE - plots are given for genes with absent signal;
if FALSE (default value) - plots are given for all the genes in the dataset.}
}
\description{
this is a function that creates 4 correlation plots (one per sample)
Library "HTqPCR" needed.
}

