% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCRdata.R
\name{qPCRdata}
\alias{qPCRdata}
\title{A qPCRdata function}
\usage{
qPCRdata(DataName, ControlGene)
}
\arguments{
\item{DataName}{= a string with the name of the data;}

\item{ControlGene}{= the name of the control gene.}
}
\description{
this is a function that makes a new qPCR object. Library "HTqPCR" needed.
}

