% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCRexprs.R
\name{qPCRexprs}
\alias{qPCRexprs}
\title{A qPCRexprs function}
\usage{
qPCRexprs(object1, object2)
}
\arguments{
\item{object1}{= first object;}

\item{object2}{= second object.}
}
\description{
this is a function that makes one qPCR class object from two qPCR objects of one kind. Library "HTqPCR" needed.
}

