\name{TaqManDat}
\docType{data}
\alias{TaqManDat}
\title{The qPCR data from GEO, "GPL4097".}
\description{It is 1044 by 16 matrix of delta Ct values, together with gene names and nondetect status.}
\usage{
data(TaqManDat)
}
\format{A qPCR data.}

\keyword{datasets}
\examples{
 data(TaqManDat)
 show(TaqManDat)
}
