% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toGetData.R
\name{toGetData}
\alias{toGetData}
\title{A toGetData function}
\usage{
toGetData(data, NDstatus, from, trtSubs)
}
\arguments{
\item{data}{= vector of qPCR signal;}

\item{NDstatus}{= vector of nondetect status;}

\item{from}{= the first record og the plate,
here it is 1, 97, 193, 289, 385, 481, 577, 673, 769, 865, 961}

\item{trtSubs}{= the first position of different treatments,
here it is 1,5,9 or 13;}
}
\description{
this is a function that subsets the data and back-calculates
Ct values, it returns vector of Ct values
}

